Transfer Dilution Updated
The function input transfer_dilution has been changed from being the total dilution across multiple dilutions to the dilution of a single transfer. A new input for number of transfers has been added to achieve the same calculation. All of the datasets and the template have also been updated to contain the relevant new parameters and have consistent formatting.